+++ 02ConvMP/src/imgutil.h
@@ -7,7 +7,7 @@ typedef std::tuple<float*,int,int> imgdata;
 namespace imgutil {
     imgdata getImageData(const std::string& fileName);
     void saveImageData(const float* pValues, int w, int h, const std::string& fileName);
-    void kernelNorm(float* kernel, int w, int h, double scale, double offset);
+    void kernelNorm(float* kernel, int w, int h, float scale, float offset);
 }
 
 #endif /* IMGUTIL_H_ */
